Fix the guessing game....
I actually cloned the repro and made the needed updates without guessing
this time.

+++ src/my_funcs.rs
@@ -1,6 +1,6 @@
 use std::cmp::Ordering;
 
-pub fn compare(input: usize, test: usize) -> Ordering {
+pub fn compare(input: u32, test: u32) -> Ordering {
     if input < test {
         Ordering::Less
     }
